+++ sdk/translation/azure-ai-translation-document/azure/ai/translation/document/_configuration.py
@@ -30,13 +30,13 @@ class DocumentTranslationClientConfiguration:  # pylint: disable=too-many-instan
      AzureKeyCredential type or a TokenCredential type. Required.
     :type credential: ~azure.core.credentials.AzureKeyCredential or
      ~azure.core.credentials.TokenCredential
-    :keyword api_version: The API version to use for this operation. Default value is "2024-05-01".
+    :keyword api_version: The API version to use for this operation. Default value is "2026-03-01".
      Note that overriding this default value may result in unsupported behavior.
     :paramtype api_version: str
     """
 
     def __init__(self, endpoint: str, credential: Union[AzureKeyCredential, "TokenCredential"], **kwargs: Any) -> None:
-        api_version: str = kwargs.pop("api_version", "2024-05-01")
+        api_version: str = kwargs.pop("api_version", "2026-03-01")
 
         if endpoint is None:
             raise ValueError("Parameter 'endpoint' must not be None.")
@@ -85,13 +85,13 @@ class SingleDocumentTranslationClientConfiguration:  # pylint: disable=too-many-
      AzureKeyCredential type or a TokenCredential type. Required.
     :type credential: ~azure.core.credentials.AzureKeyCredential or
      ~azure.core.credentials.TokenCredential
-    :keyword api_version: The API version to use for this operation. Default value is "2024-05-01".
+    :keyword api_version: The API version to use for this operation. Default value is "2026-03-01".
      Note that overriding this default value may result in unsupported behavior.
     :paramtype api_version: str
     """
 
     def __init__(self, endpoint: str, credential: Union[AzureKeyCredential, "TokenCredential"], **kwargs: Any) -> None:
-        api_version: str = kwargs.pop("api_version", "2024-05-01")
+        api_version: str = kwargs.pop("api_version", "2026-03-01")
 
         if endpoint is None:
             raise ValueError("Parameter 'endpoint' must not be None.")